XPerienceSettupWizard: offer to use WiFi to sign into Google Services
signed by: Carlos Jesús <xxx.reptar.rawrr.xxx@gmail.com>

Google setup wizard prompts for WiFi only if a data connection is not
available, and this is not the desired behavior. So let's prompt the
user whether they wish to connect to a WiFi network after selecting
their language. But we do this only once during the setup process.

Change-Id: Ifae27bac6e420217aa6a870d569b366a2273aec3